Updated CloudInformation
